requirements.txt
gradio==4.44.1

numpy==1.26.4
opencv-python==4.10.0.84
pillow==10.4.0
pyyaml==6.0.2
tqdm==4.66.5
requests==2.32.3

psutil==6.0.0
gputil==1.4.0
tabulate==0.9.0

# Real-ESRGAN from local repo
-e ./Real-ESRGAN